apps/api-auth/pyproject.toml
[tool.poetry]
name = "api-auth"
version = "0.1.0"
description = "Authentication and authorization microservice for Adyela"
authors = ["Adyela Team <dev@adyela.care>"]
readme = "README.md"
packages = [{include = "adyela_api_auth"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
gunicorn = "^21.2.0"
pydantic = {extras = ["email"], version = "^2.5.0"}
pydantic-settings = "^2.1.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
google-cloud-firestore = "^2.14.0"
google-cloud-secret-manager = "^2.17.0"
firebase-admin = "^6.4.0"
httpx = "^0.26.0"
python-dotenv = "^1.0.0"

# Observability
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
opentelemetry-exporter-gcp-trace = "^1.6.0"
prometheus-client = "^0.19.0"
structlog = "^24.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.2"
pytest-mock = "^3.12.0"
httpx = "^0.26.0"
faker = "^22.0.0"
ruff = "^0.1.11"
black = "^23.12.1"
mypy = "^1.8.0"
bandit = "^1.7.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "UP", "S", "B", "A", "C4", "DTZ", "T10", "EM", "G", "PIE", "PT", "RET", "SIM", "ARG", "PTH", "PD", "PL", "NPY", "RUF"]
ignore = ["S101"]  # Allow assert in tests

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "google.cloud.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --cov=adyela_api_auth"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["adyela_api_auth"]
omit = ["*/tests/*", "*/__pycache__/*", "*/site-packages/*"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
